+++ src/lambdas/sqs/find-actionable-events/config.ts
@@ -1,6 +1,20 @@
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export type ActionableEventsConfig = {};
+import type {
+  ActionableEventFoundTopicEndpointConfig,
+  DeadLetterQueueEndpointConfig
+} from '@model/Config';
+import {
+  readActionableEventFoundTopicConfig,
+  readDeadLetterQueueConfig,
+  readEnv
+} from '@services/common/config';
+
+export type ActionableEventsConfig = ActionableEventFoundTopicEndpointConfig &
+  DeadLetterQueueEndpointConfig;
 
 export function readActionableEventsConfig(): ActionableEventsConfig {
-  return {};
+  const env = readEnv();
+  return {
+    ...readActionableEventFoundTopicConfig(env),
+    ...readDeadLetterQueueConfig(env)
+  };
 }