+++ src/services/sns.ts
@@ -3,12 +3,14 @@ import { snsClient } from '@clients/sns';
 import { logger } from '@common/powertools';
 import type { SnsTopicConfig } from '@model/Config';
 import type { BaseEvent } from '@model/app-events/BaseEvent';
-import { extractErrorMessage } from './common/error-handling';
+import { BaseAwsMessagingService } from './common/base-aws-messaging-service';
+import { extractErrorMessage, throwError } from './common/error-handling';
 
-export class SnsService {
+export class SnsService extends BaseAwsMessagingService {
   private readonly _client: SNSClient;
   private readonly _config: SnsTopicConfig;
   private constructor(config: SnsTopicConfig) {
+    super();
     this._config = config;
     this._client = snsClient();
   }
@@ -17,40 +19,11 @@ export class SnsService {
     return new this(config);
   }
 
-  public publishEvent<TEvent extends BaseEvent>(event: TEvent): Promise<PublishCommandOutput> {
+  public publish<TEvent extends BaseEvent>(event: TEvent): Promise<PublishCommandOutput> {
     const publishCommand = new PublishCommand({
       TopicArn: this._config.topicArn,
       Message: JSON.stringify(event),
-      MessageAttributes: {
-        EventId: {
-          DataType: 'String',
-          StringValue: event.eventId
-        },
-        CorrelationId: {
-          DataType: 'String',
-          StringValue: event.correlationId
-        },
-        EventType: {
-          DataType: 'String',
-          StringValue: event.eventType
-        },
-        HappenedAt: {
-          DataType: 'String',
-          StringValue: event.happenedAt
-        },
-        UserId: {
-          DataType: 'String',
-          StringValue: event.userId
-        },
-        Idp: {
-          DataType: 'String',
-          StringValue: event.idp
-        },
-        IdpId: {
-          DataType: 'String',
-          StringValue: event.idpId
-        }
-      },
+      MessageAttributes: this.messageAttributes(event),
       MessageDeduplicationId: event.eventId,
       MessageGroupId: '1'
     });
@@ -62,11 +35,8 @@ export class SnsService {
         return result;
       },
       (error) => {
-        logger.error(
-          `Error publishing an event to SNS with id ${event.eventId}. Error: ${JSON.stringify(error)}. Extracted error: ${extractErrorMessage(error)}`
-        );
-        logger.info(`Moving on after error...`);
-        return {} as PublishCommandOutput;
+        const msg = `Error publishing an event to SNS with id ${event.eventId}. Error: ${JSON.stringify(error)}. Extracted error: ${extractErrorMessage(error)}`;
+        throwError(msg);
       }
     );
   }