+++ src/testing/utils/config.ts
@@ -6,7 +6,8 @@ import type {
   EncodeAccessJwtConfig,
   EncodeRefreshJwtConfig,
   IdpConfigs,
-  SnsTopicConfig
+  SnsTopicConfig,
+  SqsQueueConfig
 } from '@model/Config';
 import type { RefreshTokenBaseStoreConfig } from '@services/stores/refresh-token-base-store';
 import type { UserBaseStoreConfig } from '@services/stores/user-base-store';
@@ -61,8 +62,16 @@ export function setEnvUserLiveStoreConfig(config: UserLiveIndexStoreConfig): voi
   process.env.USERS_PAGE_SIZE = config.pageSize.toString();
 }
 
-export function setEnvUserCalendarFetchedConfig(config: SnsTopicConfig): void {
+export function setEnvUserCalendarFetchedTopicConfig(config: SnsTopicConfig): void {
   process.env.USER_CALENDAR_FETCHED_TOPIC_ARN = config.topicArn;
+}
+
+export function setEnvActionableEventFoundTopicConfig(config: SnsTopicConfig): void {
+  process.env.ACTIONABLE_EVENT_FOUND_ARN = config.topicArn;
+}
+
+export function setEnvDeadLetterQueueConfig(config: SqsQueueConfig): void {
+  process.env.DEAD_LETTER_QUEUE_URL = config.queueUrl;
 }
 
 export function setEnvCronRunConfig(config: CronRunConfig): void {